docs: update README and config for feature extraction caching and monitoring
- Enhanced README.md to clarify feature extraction caching behavior and added instructions for forcing a full recompute.
- Updated config/default.yaml to introduce a new parameter for monitoring minimum crisis weeks.
- Refactored src/dashboard/app.py to improve data loading and monitoring mode functionality.
- Added new utility functions in src/pipeline/run_train.py for summarizing model performance and detecting anomalies.
- Improved feature extraction logic in src/pipeline/run_features.py to streamline cache validation.
